Use float for amount in Braintree payment method
Change the `$amount` parameter from int to float to support more precise payment values. This ensures compatibility with currencies requiring fractional amounts.

+++ src/PaymentGateways/Braintree/BraintreeGateway.php
@@ -182,7 +182,7 @@ class BraintreeGateway implements PaymentGateway
         Address $billingAddress,
         string $braintreeCustomerId,
         string $braintreePaymentMethodToken,
-        int $amount,
+        float $amount,
         string|null $descriptor = null
 
     ): mixed
